Supervisor CSRF Secret has unique Type
