+++ internal/controller/supervisorconfig/generator/secret_helper.go
@@ -27,6 +27,9 @@ type SecretHelper interface {
 }
 
 const (
+	// SupervisorCSRFSigningKeySecretType is corev1.Secret.Type for the Supervisor's CSRF signing key Secret.
+	SupervisorCSRFSigningKeySecretType = "secrets.pinniped.dev/supervisor-csrf-signing-key"
+
 	// symmetricSecretType is corev1.Secret.Type of all corev1.Secret's generated by this helper.
 	symmetricSecretType = "secrets.pinniped.dev/symmetric"
 	// symmetricSecretDataKey is the corev1.Secret.Data key for the symmetric key value generated by this helper.
